# doc/openapi.yaml
openapi: 3.0.3
info:
  title: QuickEvent exchange httpd - OpenAPI 3.0
  description: |-
    This is QuickEvent exchange server based on the OpenAPI 3.0 specification.  
    
    Some useful links:
    - [The Quick Event repository](https://github.com/Quick-Event/quickbox)
  version: 0.0.1
paths:
  /:
    get:
      tags:
        - web
      summary: Get event list html page
      operationId: index
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: string
  /event/{eventId}:
    get:
      tags:
        - web
      summary: Get event info html page
      operationId: eventInfo
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/EventId'
      responses:
        '200':
          description: Success
          content:
            text/json:
              schema:
                $ref: '#/components/schemas/EventInfo'
        '404':
          description: Invalid event key
  /api/event/{eventId}:
    get:
      tags:
        - API
      summary: Get event record
      operationId: get_event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/EventId'
      responses:
        '200':
          description: Success
          content:
            text/json:
              schema:
                $ref: '#/components/schemas/EventInfo'
        '404':
          description: Invalid event ID
  /api/event:
    post:
      tags:
        - API
      summary: Register event
      operationId: api_register_event
      parameters:
        - name: registerToken
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInfo'
        required: true
      responses:
        '200':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterEventResponse'
        '400':
          description: Invalid register token
  /api/event/{eventId}/oc:
    post:
      tags:
        - API
      summary: Add new O-Checklist change set
      # description: Add a new pet to the store
      operationId: add_oc_change_set
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/EventId'
        - name: apiKey
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/ApiKey'
      requestBody:
        # description: Add new recor set
        content:
          application/yaml:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/OChecklistRunner'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Invalid input
        '422':
          description: Validation exception
    # get:
    #   tags:
    #     - API
    #   summary: Read change set list
    #   # description: Add a new pet to the store
    #   operationId: listOChecklistChangeSets
    #   parameters:
    #   - name: eventId
    #     in: path
    #     required: true
    #     schema:
    #       $ref: '#/components/schemas/EventId'
    #   - name: idgt
    #     in: query
    #     description: only change-sets with ID greater will be returned
    #     required: false
    #     schema:
    #       $ref: '#/components/schemas/ChangeSetId'
    #   - name: eventKey
    #     in: header
    #     required: true
    #     schema:
    #       $ref: '#/components/schemas/EventKey'
    #   responses:
    #     '200':
    #       description: Successful operation
    #       content:
    #         application/json:
    #           schema:
    #             type: array
    #             items: 
    #               $ref: '#/components/schemas/OChecklistChangeSet'
    #     '404':
    #       description: Invalid event key oe event Id   
  /api/event/{eventId}/qe/chng/in:
    get:
      tags:
        - API
      summary: Read change set list
      operationId: get_in_changes
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/EventId'
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            minimum: 0
      responses:
        '200':
          description: Table of changes with index greater or equal offset, may be paginated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QEChangeRecord'
        '404':
          description: Invalid event key
  /api/event/{eventId}/qe/startListIOF30Xml:
    get:
      tags:
        - API
      summary: Read actual start list
      # description: Add a new pet to the store
      operationId: get_start_list_iofxml3
      parameters:
        - name: eventId
          in: path
          description: Event key generated by Quick Event
          required: true
          schema:
            $ref: '#/components/schemas/EventId'
      responses:
        '200':
          description: Successful operation
          content:
            application/xml:
              schema:
                type: string
        '404':
          description: Invalid event key
    put:
      tags:
        - API
      summary: Write actual start list
      # description: Add a new pet to the store
      operationId: set_start_list_iofxml3
      parameters:
        - name: eventId
          in: path
          description: Event key generated by Quick Event
          required: true
          schema:
            $ref: '#/components/schemas/EventId'
        - name: apiKey
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/ApiKey'
      requestBody:
        content:
          application/xml:
            schema:
              type: string
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/text:
              schema:
                type: string
        '404':
          description: Invalid event key or event id
components:
  schemas:
    RegisterEventResponse:
      properties:
        eventId:
          type: integer
        apiKey:
          $ref: '#/components/schemas/ApiKey'
    EventInfo:
      properties:
        dbName:
          type: string
        name:
          type: string
        date:
          type: string
        place:
          type: string
    SiId:
      type: integer
      format: int64
      description: SportIdend card number
      example: 987456
    RunId:
      type: integer
      format: int64
      description: QE run ID
      example: 987456
    EventId:
      type: string
      description: Event id generated by QX server
    ApiKey:
      type: string
      description: Unique token to authorize write to event data
    OChecklistRunner:
      properties:
        Id:
          type: string
          format: int64
          example: 1234
        StartStatus:
          type: string
          enum:
            - Started OK
            - Did not start (DNS)
            - Late start
        Card:
          $ref: '#/components/schemas/SiId'
        ClassName:
          type: string
          # format: int64
          example: H21
        Name:
          type: string
          # format: int64
          example: John Doe
        StartTime:
          $ref: '#/components/schemas/Time'
        Comment:
          type: string
    OChecklistChangeSet:
      properties:
        id:
          $ref: '#/components/schemas/ChangeSetId'
        records:
          type: array
          items:
            $ref: '#/components/schemas/OChecklistRunner'
    ChangeSetId:
      type: integer
      example: 123
    Time:
      type: string
      pattern: "^[0-2][0-9]:[0-5][0-9]:[0-5][0-9](.[0-9]+)?$"
      example: 17:18:59.123
    DateTime:
      type: string
      pattern: "^20[0-9]{2}-[0-1][0-9]-[0-3][0-9]T[0-2][0-9]:[0-5][0-9]:[0-5][0-9](.[0-9]+)?$"
      example: 17:18:59.123
    QEChangeRecord:
      properties:
        chngId:
          $ref: '#/components/schemas/ChangeSetId'
        table:
          type: string
          enum:
            - runs # QERunsRecord
            - radio # QERadioRecord
          
          # corresponding table record properties follows 

    QERunsRecord:
      properties:
        siId:
          $ref: '#/components/schemas/SiId'
        runId:
          $ref: '#/components/schemas/RunId'
        checkTime:
          $ref: '#/components/schemas/EventId'
        comment:
          type: string
    QERadioRecord:
      properties:
        siId:
          $ref: '#/components/schemas/SiId'
        punchTime:
          $ref: '#/components/schemas/EventId'